+++ A_FIT_DAY/urls.py
@@ -27,7 +27,6 @@ urlpatterns = [
     path("api/posts/", include("posts.urls")),
     path("api/articles/", include("articles.urls")),
     path("api/chat_gpt/", include("chatgpt.urls")),
-    path('crawler/', include('crawler.urls')),
 ]
 
 # 미디어 파일 제공 설정 추가
