+++ articles/urls.py
@@ -1,3 +1,7 @@
 from django.urls import path
+from .views import CrawlerNewsList, CrawlerHealthChosun
 
-urlpatterns = []
+urlpatterns = [
+    path('news/', CrawlerNewsList.as_view(), name='news_list'),
+    path('Chosun/', CrawlerHealthChosun.as_view(), name='news_list'),
+]
